requirements.txt
gradio
pytorch-transformers==1.0.0
Flask==1.1.1
Flask-Cors==3.0.8
opencv-python
torch
torchvision
tensorflow==2.0.0b1
git+https://github.com/huggingface/transformers.git